MAILBOX-359 Gathering non related event DTOs into one class
