+++ mailbox/api/src/main/java/org/apache/james/mailbox/FlagsBuilder.java
@@ -61,11 +61,6 @@ public class FlagsBuilder {
         return this;
     }
 
-    public FlagsBuilder merge(FlagsBuilder flagsBuilder) {
-        internalFlags.add(flagsBuilder.internalFlags);
-        return this;
-    }
-
     public Flags build() {
         return new Flags(internalFlags);
     }
